requirements.txt

pandas
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
#python-bioc
pyYaml
tqdm
ensure
joblib
types-pyYaml
#box
pyYaml

#Flask
#Flask-Config
zenml
-e .

